aitune: backport byte-match-proven fix — GetOneWay reads +61 bytes of BWorldSm_slices
The matching decomp byte-matched AITune_GetOneWay, proving the high-track branch
loads byte +61 of BWorldSm_slices (lbu v0,61). The run-tree used typed-pointer
arithmetic `BWorldSm_slices + 0x3d` = +0x3d*sizeof(Trk_NewSlice=32) = +1952 bytes
(wrong address). Cast to (int) for byte offset. Compiles clean under ccpsx 4.3.

+++ game/common/AITUNE.cpp
@@ -34,7 +34,7 @@ void AITune_CleanUp2(void);
 int AITune_GetOneWay(void)
 {
   if (0xb < GameSetup_gData.track) {
-    return (u_int)(*(u_char *)(BWorldSm_slices + 0x3d) >> 4 == 0);
+    return (u_int)(*(u_char *)((int)BWorldSm_slices + 0x3d) >> 4 == 0);   /* byte-match backport: +61 bytes, not +0x3d*sizeof(Trk_NewSlice=32) */
   }
   return AITune_trackInfo[GameSetup_gData.track].oneWay;
 }
